fixes #3902 - [Openstack] allow selection of internal networks
This patch allows the user to select the internal network when deploying
a new instance in openstack.

In Grizlly, this was done automaticilly, with a nasty side effect, that
all internal networks would get assigned (didnt see that as i had only
one internal network when i was testing), starting in Havana, this is no
longer possible, and the vm creation is rejected.

as a bonus, we now retain the floating ip selection if the form had an
error.