[UI] smart param order priority can now be set with up/down arrows
this introduce an inital experiement with reactjs to manipulate the
existing form, without interaption to the rest of the app.

Since javascript generates the textarea input, and we wish that its
part of the original form submit, it is required to keep the exact
html input name, therefor this patch introduce an extension to rails
form builder, to allow accessing the html name attribute directly.